+++ services/api-server/src/simcore_service_api_server/models/custom_errors.py
@@ -1,12 +1,7 @@
-import logging
-
-from fastapi import Request, status
-from starlette.responses import JSONResponse
-
-_logger = logging.getLogger(__name__)
+from .errors import ApiServerBaseError
 
 
-class CustomBaseError(Exception):
+class CustomBaseError(ApiServerBaseError):
     pass
 
 
@@ -20,14 +15,3 @@ class MissingWalletError(CustomBaseError):
 
 class ApplicationSetupError(CustomBaseError):
     pass
-
-
-async def custom_error_handler(_: Request, exc: CustomBaseError):
-    if isinstance(exc, InsufficientCreditsError):
-        return JSONResponse(
-            status_code=status.HTTP_402_PAYMENT_REQUIRED, content=f"{exc}"
-        )
-    if isinstance(exc, MissingWalletError):
-        return JSONResponse(
-            status_code=status.HTTP_424_FAILED_DEPENDENCY, content=f"{exc}"
-        )
